Compiles & runs, but disp double values inaccurate
